Add all files/dirs under /etc/etckeeper
As we're about to start managing the /etc/etckeeper/*.d dirs to
alter the behaviour, its good for us to completely manage these directories
ourselves.

We can then mark the directories in /etc/etckeeper as 'clean: True', and
have salt automatically purge files created under here manually.

Considering we're using this as a layer of defense in our security stack,
this is worth the effort to maintain.

NB: Unfortunately salt does not allow for simply saying 'this file is ok', so
    we must override the package management of these files.